Alpha version 0.3.2
- introduced a somewhat user interface
- checker module now shows changes
- remover module checks whether there are any files to remove first

+++ remover_module.py
@@ -5,22 +5,30 @@ import os
 # existing searches checker
 def remover():
     print("\nRemover initiated")
-    print("\nWhich search would you like to remove?")
-    fileName = input("File Name (without .csv extension): ")
-    fileName = fileName + ".csv"
-
     # open file to read lines
     with open("csvFilesIndex.txt", mode="r") as cFi:
         lines = cFi.readlines()
         cFi.close()
 
-    # rewrite file without removed filename
-    with open("csvFilesIndex.txt", "w") as cFi:
+    if len(lines) != 0:
+        print("Currently active files:")
         for line in lines:
-            if line.strip("\n") != fileName:
-                cFi.write(line)
-    try:
-        os.remove(fileName)
-        print("File removed successfully")
-    except:
-        print("File inexistent")
+            print(line.strip("\n"))
+
+        print("\nWhich search would you like to remove?")
+        fileName = input("File Name (without .csv extension): ")
+        fileName = fileName + ".csv"
+
+        # rewrite file without removed filename
+        with open("csvFilesIndex.txt", "w") as cFi:
+            for line in lines:
+                if line.strip("\n") != fileName:
+                    cFi.write(line)
+        try:
+            os.remove(fileName)
+            print("File removed successfully")
+        except:
+            print("File inexistent")
+
+    else:
+        print("No files to remove")
